docmind/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.1
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
openai==1.30.5
langchain==0.2.5
langchain-openai==0.1.9
langchain-community==0.2.5
faiss-cpu==1.8.0
pinecone-client==4.0.0
pymupdf==1.24.5
tiktoken==0.7.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.7
pytest==8.2.2
pytest-asyncio==0.23.7
